requirements.txt
python-docx
transformers
langchain
torch
pdfplumber
protobuf
tiktoken
blobfile
sentencepiece